Remove HFDM references + few renaming
