+++ experimental/PropertyDDS/examples/property_inspector/src/index.html
@@ -7,7 +7,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2280%22>💧</text></svg>">
-    <title>Fluid Hello World</title>
+    <title>Property Inspector</title>
 </head>
 <body>
     <div style="height: 100%" id="root"></div>
