+++ experimental/PropertyDDS/packages/property-binder/README.md
@@ -1,7 +1,7 @@
 # @fluid-experimental/property-binder
 
 ## Overview
-Forge Application Framework Databinder provides functionality to easily bind data stored in HFDM to business logic.
+Databinder provides functionality to easily bind data stored in PropertyTree to business logic.
 
 ## Installation
