+++ experimental/PropertyDDS/packages/property-inspector-table/src/index.ts
@@ -7,7 +7,7 @@ import InspectorTableIcons from '../assets/icons/SVGStoreIcons';
 export { InspectorTableIcons };
 export * from './InspectorTable';
 export * from './InspectorTableTypes';
-export * from './HFDMDataCreationHandlers';
+export * from './PropertyDataCreationHandlers';
 export * from './icons';
 export * from './ModalManager';
 export * from './ModalRoot';
